💄 🍱 이미지 화질개선 및 스타일 추가 분리
