+++ client/components/ClickButton.js
@@ -1,6 +1,5 @@
 import React from "react";
 import { StyleSheet, TouchableOpacity, Text, View } from "react-native";
-import { buttonStyle } from "../screens/Styles";
 
 const ClickButton = (props) => {
   return (
@@ -13,3 +12,26 @@ const ClickButton = (props) => {
 };
 
 export default ClickButton;
+
+export const buttonStyle = StyleSheet.create({
+  container: {
+    //flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  button: {
+    width: 107,
+    height: 31,
+    borderRadius: 38.75,
+    backgroundColor: "#EC0A0A",
+    justifyContent: "center",
+  },
+  buttonText: {
+    textAlign: "center",
+    color: "#F9F9F9",
+    fontSize: 15.5,
+    fontStyle: "normal",
+    fontWeight: "600",
+    letterSpacing: 0.31,
+  },
+});
