+++ client/screens/Home.js
@@ -243,13 +243,10 @@ const style = StyleSheet.create({
     fontSize: 11,
     fontStyle: "normal",
     fontWeight: "600",
-    // alignItems: "center",
     width: 85,
-    // justifyContent: "center",
     marginTop: 8,
-    // marginLeft: "auto",
-    // marginRight: "auto",
     textAlign: "center",
+    marginLeft: 9,
   },
   menuImage: {
     width: 77,
