+++ client/screens/SignIn.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Text, TextInput, View, Button, Image } from "react-native";
+import { Text, TextInput, View, Button, Image, StyleSheet } from "react-native";
 import ClickButton from "../components/ClickButton";
 import logo from "../assets/icons/logo.png";
 // 로고는 추후에 수정 예정
